+++ example/components/EVMConnectButton.tsx
@@ -1,28 +1,50 @@
 import { useState } from "react";
 
-import { BlockchainTypes, CrossmintEVMWalletAdapter } from "@crossmint/connect";
+import { BlockchainTypes, CrossmintEVMWalletAdapter, CrossmintEnvironment } from "@crossmint/connect";
 
 export default function EVMConnectButton() {
     const [address, setAddress] = useState<string | undefined>(undefined);
+    const [signedMessage, setSignedMessage] = useState<string | undefined>(undefined);
+    const [crossmintConnect, setCrossmintConnect] = useState<CrossmintEVMWalletAdapter | undefined>(undefined)
 
-    async function connectToCrossmint() {
-        // Initialize the Crossmint embed.
-        const _crossmintEmbed = new CrossmintEVMWalletAdapter({
-            chain: BlockchainTypes.ETHEREUM, // BlockchainTypes.ETHEREUM || BlockchainTypes.POLYGON. For solana use BlockchainTypes.SOLANA
+    async function handleClick() {
+        // prompt user to trust your app
+        const crossmintConnect = new CrossmintEVMWalletAdapter({
+            chain: BlockchainTypes.POLYGON,
+            environment: CrossmintEnvironment.LOCAL,
+            maxTimeAutoConnectMs: 100000
         });
+        const _address = await crossmintConnect.connect();
 
-        // Ask the user to sign in and give access to their publicKey
-        const address = await _crossmintEmbed.connect();
 
-        // If the user successfully connects to Crossmint, the address will be returned.
-        if (address) {
-            setAddress(address);
-        }
+        // store the result in react state
+        setAddress(_address);
+        setCrossmintConnect(crossmintConnect)
     }
 
+    const signMessage = async () => {
+        const message = "Hello, world!";
+        if (!crossmintConnect) {
+            throw new Error('not set')
+        }
+        // Sign the message
+        const signature = await crossmintConnect.signMessage(message);
+        setSignedMessage(signature);
+    }
+
+    const connected = address != null;
+
+    // If connected, displays their address, else displays "Connect"
     return (
-        <button onClick={connectToCrossmint} className="px-6 py-2 font-semibold text-black bg-white rounded-md">
-            {address ? address.slice(0, 6) + "..." : "Connect"}
-        </button>
+        <>
+            <button onClick={handleClick} disabled={connected}>
+                {connected ? `${address.slice(0, 6)}...` : "Connect"}
+            </button>
+            <div>
+                <button onClick={signMessage}>SignMessage</button>
+                {signedMessage && <div>{signedMessage}</div>}
+            </div>
+        </>
     );
+
 }
